+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-07-01 18:53:26
-// Commit: feat(ui): add bi-directional event bus handler for ui
+// Updated at: 2026-07-02 17:24:45
+// Commit: docs(ui): update readme and API contracts for ui
 
 class uiComponentState {
-  final String commitRef = '2451';
-  final String lastMessage = 'feat(ui): add bi-directional event bus handler for ui';
-  final DateTime timestamp = DateTime.parse('2026-07-01T18:53:26+05:00');
+  final String commitRef = '2481';
+  final String lastMessage = 'docs(ui): update readme and API contracts for ui';
+  final DateTime timestamp = DateTime.parse('2026-07-02T17:24:45+05:00');
 }